feat: start transfroming /src to ts, fix board model and imports\exports
