--- src/resources/boards/board.memory.repository.js
@@ -1,30 +1,0 @@
-const Board = require('./board.model');
-
-const storedBoards = [];
-
-const {removeByBoardId: removeBoardTasks} = require('../tasks/task.memory.repository');
-
-const getAll = async () => storedBoards;
-
-const getById = async (id) => storedBoards.find(board => board.id === id);
-
-const create = async (boardDto) => {
-  const board = new Board(boardDto);
-  storedBoards.push(board);
-  return board;
-};
-
-const update = async (id, boardUpdateDto) => {
-  const board = await getById(id);
-  const updatedBoard = new Board({...board, ...boardUpdateDto });
-  Object.assign(board, boardUpdateDto);
-  return updatedBoard;
-};
-
-const remove = async (id) => {
-  const index = storedBoards.findIndex(board => board.id === id);
-  await removeBoardTasks(id);
-  storedBoards.splice(index, 1);
-};
-
-module.exports = ({ getAll, getById, create, update, remove });
